[DAD-860] refactor: BoardRepository 파일 위치 변경
