+++ src/test/java/com/forever/dadamda/service/BoardServiceTest.java
@@ -6,7 +6,7 @@ import com.forever.dadamda.dto.board.CreateBoardRequest;
 import com.forever.dadamda.entity.board.Board;
 import com.forever.dadamda.entity.board.TAG;
 import com.forever.dadamda.entity.user.User;
-import com.forever.dadamda.repository.BoardRepository;
+import com.forever.dadamda.repository.board.BoardRepository;
 import com.forever.dadamda.repository.UserRepository;
 import java.time.LocalDateTime;
 import java.util.Optional;
